app/requirements.txt
aiogram
elasticsearch==8.15.1
anthropic
sqlalchemy
asyncpg
python-dotenv
